refactor: extract TruncatedPath component
Extract shared path truncation rendering into a
reusable TruncatedPath component. Used by both file
tree entries and the output path footer.
